# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yamovo_contentx_internal_models.ContentEntry:
    properties:
      content_type:
        $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentType'
      content_type_id:
        type: integer
      created_at:
        type: string
      created_by_id:
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/github_com_yamovo_contentx_internal_models.JSONMap'
        description: field values as JSON
      document_id:
        description: UUID
        type: string
      id:
        type: integer
      locale:
        description: |-
          i18n: Locale is the BCP-47 language tag. TranslationGroupID links
          translations of the same logical entry across locales.
        type: string
      published_at:
        type: string
      status:
        description: draft, published
        type: string
      translation_group_id:
        type: integer
      updated_at:
        type: string
      updated_by_id:
        type: integer
    type: object
...
  github_com_yamovo_contentx_internal_models.ContentType:
    properties:
      created_at:
        type: string
      description:
        type: string
      draft_publish:
        description: supports draft/publish
        type: boolean
      entry_count:
        type: integer
      fields:
        items:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentField'
        type: array
      id:
        type: integer
      is_single:
        description: single vs collection
        type: boolean
      name:
        description: display name
        type: string
      uid:
        description: e.g. "product", "event"
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_yamovo_contentx_internal_models.JSONMap:
    additionalProperties: true
    type: object
  github_com_yamovo_contentx_internal_models.ListResponse:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_yamovo_contentx_internal_services.CreateContentTypeRequest:
    properties:
      description:
        type: string
      draft_publish:
        type: boolean
      fields:
        items:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateFieldRequest'
        minItems: 1
        type: array
      is_single:
        type: boolean
      name:
        maxLength: 128
        type: string
      uid:
        maxLength: 64
        type: string
    required:
    - fields
    - name
    - uid
    type: object
  github_com_yamovo_contentx_internal_services.CreateEntryRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      locale:
        description: 'i18n: BCP-47 tag, defaults to "en"'
        type: string
      status:
        description: draft (default) or published
        type: string
    required:
    - data
    type: object
...
  github_com_yamovo_contentx_internal_services.UpdateEntryRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      status:
        type: string
    type: object
...
  internal_handlers.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      err_code:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/internal_handlers.Meta'
    type: object
  internal_handlers.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: High-performance Headless CMS - API-first content platform
  title: ContentX API
  version: "1.0"
paths:
...
  /content-types:
    get:
      description: Returns all user-defined content types with field definitions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentType'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List content types
      tags:
      - Content Types
    post:
      consumes:
      - application/json
      description: Create a new content type with field definitions
      parameters:
      - description: Content type definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateContentTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create content type
      tags:
      - Content Types
  /content-types/{uid}:
    delete:
      description: Deletes a content type and all its entries
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete content type
      tags:
      - Content Types
    get:
      description: Returns a content type with its field definitions
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentType'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get content type
      tags:
      - Content Types
  /content/{uid}:
    get:
      description: Returns paginated entries of a content type
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by status
        enum:
        - draft
        - published
        in: query
        name: status
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      - default: newest
        description: Sort order
        enum:
        - newest
        - oldest
        - updated
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List entries
      tags:
      - Content Entries
    post:
      consumes:
      - application/json
      description: Create a new entry for a content type
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Entry data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentEntry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create entry
      tags:
      - Content Entries
  /content/{uid}/{documentId}:
    delete:
      description: Delete an entry by document ID
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete entry
      tags:
      - Content Entries
    get:
      description: Returns a single entry by document ID
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentEntry'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get entry
      tags:
      - Content Entries
    put:
      consumes:
      - application/json
      description: Update an existing entry
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.UpdateEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentEntry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update entry
      tags:
      - Content Entries
  /content/{uid}/{documentId}/publish:
    post:
      description: Publish a draft entry
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentEntry'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Publish entry
      tags:
      - Content Entries
  /content/{uid}/{documentId}/translations:
    get:
      description: Returns sibling translations of an entry (same translation group)
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List entry translations
      tags:
      - Content Entries
    post:
      consumes:
      - application/json
      description: Creates a new entry as a translation of an existing one. Requires
        content.create permission.
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      - description: Target locale (BCP-47 tag, e.g. en, zh)
        in: query
        name: locale
        required: true
        type: string
      - description: Entry translation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentEntry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create entry translation
      tags:
      - Content Entries
  /content/{uid}/{documentId}/unpublish:
    post:
      description: Revert a published entry to draft
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ContentEntry'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Unpublish entry
      tags:
      - Content Entries
  /content/{uid}/export:
    get:
      description: Export all entries as JSON
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Export entries
      tags:
      - Content Entries
  /content/{uid}/import:
    post:
      consumes:
      - application/json
      description: Import entries from JSON data
      parameters:
      - description: Content Type UID
        in: path
        name: uid
        required: true
        type: string
      - description: JSON data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  properties:
                    imported:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Import entries
      tags:
      - Content Entries
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"